requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.1
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
mccabe==0.6.1
psycopg2
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
gunicorn
